functions-datapipeline/function-bq-insert/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.13.*